backend/requirements.txt
fastapi
motor
pymongo
uvicorn
requests
python-dotenv
httpx